finished cheesed redo of schematic and layout for v1.0 production board
This pass is the first at something that will eventually
look pretty much like the production PCB. Changes include
removing all the capacitors as gratuitous, removing the
gratuitous touchpad resistor, moving the battery to the back
and enlarging the tabs for hand soldering, moving to the new
LEDs, and changing to the USB Debug Device compatible debug
connector.

Changes for the next version include better re-layout with
less vias (maybe), removing the terrible debug connector as
too huge, and perhaps removing more components if
experimentation indicates.

The board id text is currently tiny and buried under the
battery once it's soldered on. Not much can be done, really.